SHMRK-1339 Added success message after requesting an appointment
Added success message after requesting an appointment. Replaced the use
of the default android dialog by pendent/sweet-alert-dialog. Added a
visibility mechanism for Tablet classes to know when an activity is in
the ForeGround. So basically in any Tablet class there an isVisible
property that let the activity check ForeGround status. I’ll add the
same mechanism to the phone classes tomorrow.